created submission_file_util file and moved enum functions there, edited references to those functions
