+++ src/cromshell/utilities/workflow_id_utils.py
@@ -1,7 +1,7 @@
 import csv
 import logging
 
-from cromshell.utilities import cromshellconfig, io_utils
+from cromshell.utilities import cromshellconfig, io_utils, submissions_file_utils
 
 LOGGER = logging.getLogger(__name__)
 
@@ -62,7 +62,9 @@ def obtain_workflow_id_using_digit(relative_id: int, submission_file_path: str) 
         relative_id if relative_id < 0 else relative_id - 1
     )  # Subtracting 1 to account for row shift when dealing with list index
 
-    return mycsv[row_index][cromshellconfig.ImmutableSubmissionFileHeader.Run_ID.value]
+    return mycsv[row_index][
+        submissions_file_utils.ImmutableSubmissionFileHeader.Run_ID.value
+    ]
 
 
 def obtain_workflow_id_using_alias(alias_name: str, submission_file_path: str) -> str:
@@ -77,10 +79,12 @@ def obtain_workflow_id_using_alias(alias_name: str, submission_file_path: str) -
         reader = csv.DictReader(csv_file, delimiter="\t")
         for row in reader:
             if (
-                row[cromshellconfig.MutableSubmissionFileHeader.Alias.value]
+                row[submissions_file_utils.MutableSubmissionFileHeader.Alias.value]
                 == alias_name
             ):
-                return row[cromshellconfig.ImmutableSubmissionFileHeader.Run_ID.value]
+                return row[
+                    submissions_file_utils.ImmutableSubmissionFileHeader.Run_ID.value
+                ]
 
         LOGGER.error(
             "Unable to find alias '%s' in submission file '%s'",
@@ -101,7 +105,7 @@ def workflow_id_exists(workflow_id: str, submission_file) -> bool:
         reader = csv.DictReader(csv_file, delimiter="\t")
         for row in reader:
             if (
-                row[cromshellconfig.ImmutableSubmissionFileHeader.Run_ID.value]
+                row[submissions_file_utils.ImmutableSubmissionFileHeader.Run_ID.value]
                 == workflow_id
             ):
                 return True
