+++ tests/integration/test_alias.py
@@ -4,7 +4,7 @@ import shutil
 
 import pytest
 
-from cromshell.utilities import cromshellconfig
+from cromshell.utilities import submissions_file_utils as sfu
 from tests.integration import utility_test_functions
 
 
@@ -17,23 +17,16 @@ def assert_workflow_id_matches_alias(
     with open(local_workflow_database_tsv, "r") as csv_file:
         reader = csv.DictReader(csv_file, delimiter="\t")
         for row in reader:
-            if (
-                row[cromshellconfig.ImmutableSubmissionFileHeader.Run_ID.value]
-                == workflow_id
-            ):
+            if row[sfu.ImmutableSubmissionFileHeader.Run_ID.value] == workflow_id:
                 #  If alias_name is "" (which is when we want to remove an alias) assert
                 #  that the alias for the workflow id returns None. Else (when adding
                 #  and alias) assert that the alias for the workflow id returns
                 #  the matching alias name.
                 if alias_name == "":
-                    assert (
-                        row[cromshellconfig.MutableSubmissionFileHeader.Alias.value]
-                        is None
-                    )
+                    assert row[sfu.MutableSubmissionFileHeader.Alias.value] is None
                 else:
                     assert (
-                        row[cromshellconfig.MutableSubmissionFileHeader.Alias.value]
-                        == alias_name
+                        row[sfu.MutableSubmissionFileHeader.Alias.value] == alias_name
                     )
